requirements.txt
absl-py==0.2.2
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.13.0
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.3
Pillow==5.2.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
